Update SLY_Assistant.user.js
Bug fix:
When a planet or starbase was in the cache and a the entry was requested again (because of an expired entry or the parameter "getLive"), the cache entry wasn't overwritten. Instead the new entry was appended to the array. But when reading from the cache, it will always get the first entry that was found, so the new entry wasn't never used. Fixed by: The cached entry is now overwritten with the new values.